Add Fallible.Void type to specify a return type which returns nothing but can contain an error.
Fallible.Void can only be created using the Fallible.Return property.
