+++ Fallible/Void.cs
@@ -1,0 +1,6 @@
+namespace Fallible;
+
+public class Void
+{
+    internal Void() { }
+}
